compiler type hint != casting

+++ src/optimization.jl
@@ -176,11 +176,11 @@ function optimization_engine(
 end
 
 """
-    nan2inf(x::Float64)::Float64
+    nan2inf(x::Float64)
 
 Turn NaNs into Inf
 """
-function nan2inf(x::Float64)::Float64
+function nan2inf(x::Float64)
     if isnan(x)
         return Inf
     else
